requirements.txt
# Core dependencies
alpaca-trade-api>=3.1.1
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data handling and storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
pyarrow>=12.0.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
quantstats>=0.0.62

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
websockets>=11.0.0

# Configuration and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Async and concurrency
aiohttp>=3.8.0
aiofiles>=23.1.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0